watchlist_rating_implementation

+++ src/main/java/com/social/horror_pool/service/WatchlistService.java
@@ -26,5 +26,5 @@ public interface WatchlistService {
 
     WatchlistAllResponse getAllPublicWatchlists(Integer pageNumber, Integer pageSize, String order);
 
-    WatchlistDTO rateWatchlist(Long watchlistId, Long rating);
+    WatchlistDTO rateWatchlist(Long watchlistId, double rating);
 }
